refactor: move Query related components
