+++ src/lib/components/queries/QueryDisplay.svelte
