+++ src/lib/components/queries/QueryResult.svelte
